TUTORIAL FILES for HTML and CSS coding.
The HTML file works alongside the annexed image (puppytestimage) and the "testpage2". Read the file, and attempt to familiarize yourself with the mechanics of coding with HTML.

The CSS file works alongside the annexed image (Test Background) as well and "index2.html". It is important to read it, to familiarize yourself with the mechanics of coding CSS and HTML.

Both files are filled with comments, explaining how they (and consequently, key mechanics of HTML and CSS coding) work. Everything is in english, befittingly to our standard.

DO NOT DELETE OR NOT DOWNLOAD ANY OF THE FILES IN THIS PACKAGE. 

INDIVIDUAL DOWNLOAD WILL COMPROMISE THE FUNCTIONING OF EITHER - DOWNLOAD EVERYTHING.